Changes after CR
