+++ fplus-lib/src/external_services/mod.rs
@@ -1,3 +1,4 @@
 pub mod blockchain;
+pub mod dmob;
 pub mod filecoin;
 pub mod github;
