+++ fplus-lib/src/models/mod.rs
@@ -1,1 +1,2 @@
+pub mod dmob;
 pub mod filecoin;
